refactor: extract bookmark sync logic into dedicated hook
Extract complex bookmark synchronization logic from page.tsx into
reusable useBookmarkSync hook. This improves separation of concerns
and follows existing hook patterns in the codebase.

- Create src/hooks/use-bookmark-sync.ts with bookmark sync logic
- Hook handles cross-tab bookmark synchronization
- Extract bookmark event listeners and state sync logic
- Follow existing hook patterns and TypeScript interfaces
- Update page.tsx to use the new hook
- Preserve exact same bookmark sync functionality
